rename and convert files to ts
