+++ packages/mermaid/src/diagram-api/diagram-orchestration.ts
@@ -16,7 +16,7 @@ import stateV2 from '../diagrams/state/stateDetector-V2.js';
 import journey from '../diagrams/user-journey/journeyDetector.js';
 import errorDiagram from '../diagrams/error/errorDiagram.js';
 import flowchartElk from '../diagrams/flowchart/elk/detector.js';
-import timeline from '../diagrams/timeline/detector.js';
+import timeline from '../diagrams/timeline/timelineDetector.js';
 import mindmap from '../diagrams/mindmap/detector.js';
 import sankey from '../diagrams/sankey/sankeyDetector.js';
 import { registerLazyLoadedDiagrams } from './detectType.js';
