+++ packages/mermaid/src/diagrams/timeline/timelineDetector.ts
@@ -11,7 +11,7 @@ const detector: DiagramDetector = (txt) => {
 };
 
 const loader: DiagramLoader = async () => {
-  const { diagram } = await import('./timeline-definition.js');
+  const { diagram } = await import('./timelineDiagram.js');
   return { id, diagram };
 };
