+++ packages/mermaid/src/diagrams/timeline/timelineDiagram.ts
@@ -2,7 +2,7 @@
 import parser from './parser/timeline.jison';
 import * as db from './timelineDb.js';
 import renderer from './timelineRenderer.js';
-import styles from './styles.js';
+import styles from './timelineStyles.js';
 
 export const diagram = {
   db,
